+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,14 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - Unreleased
+
+### Fixed
+
+- When a tilde expansion produces a directory name that ends with a slash and
+  the expansion is followed by a slash, the trailing slash in the directory name
+  is now removed to maintain the correct number of slashes.
+
 ## [0.4.1] - 2025-05-03
 
 ### Fixed
@@ -182,6 +190,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of the shell
 
+[0.4.2]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.2
 [0.4.1]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.1
 [0.4.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.0
 [0.3.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.3.0